+++ src/api/products.ts
@@ -1,6 +1,6 @@
-import { Phones } from '../types/Phone';
+import { Phones } from '../types/Phones';
 import { client } from '../services/fetchClient';
-import { PhoneDetails } from '../types/PhoneDetail';
+import { PhoneDetails } from '../types/PhoneDetails';
 
 export const getProducts = (page: number, perPage: number, orderBy: string, order: string) => {
   let orderByString = '';
