+++ src/components/Card/Card.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import './card.scss';
 import { ButtonFavs } from '../utils/ButtonFavs/ButtonFavs';
 import { ButtonDefault } from '../utils/ButtonDefault/ButtonDefault';
-import { Phone } from '../../types/Phone';
+import { Phone } from '../../types/Phones';
 
 type CardProps = {
   item: Phone;
