+++ src/types/PhoneDetails.ts
