+++ src/types/Phones.ts
@@ -1,0 +1,20 @@
+export type Phone = {
+  id: number;
+  category: string;
+  phoneId: string;
+  itemId: string;
+  name: string;
+  fullPrice: number;
+  price: number;
+  screen: string;
+  capacity: string;
+  color: string;
+  ram: string;
+  year: number;
+  image: string;
+};
+
+export type Phones = {
+  count: number;
+  rows: Phone[];
+};
